requirements.txt
flask
PyJWT==2.7.0